+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- installment data to IssuingPurchase resource
 
 ## [0.12.0] - 2024-07-01
 ### Added